refactor(api): extract shared iTunes image utility
Extract duplicated extractItunesImage function into shared utility
to reduce code duplication between rss-fetcher and subscriptions router.

- Create feed-utils.ts with extractItunesImage helper
- Remove duplicate function from rss-fetcher.ts
- Remove duplicate function from subscriptions.ts
- Both modules now import from shared utility
